pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "movie_colorbar/__init__.py"

[tool.hatch.build.targets.sdist]
exclude = [
  "/.github",
]

[tool.hatch.build.targets.wheel]
packages = ["movie_colorbar"]

[project]
name = "movie_colorbar"
readme = "README.md"
description = "Turn a video into a colorbar"
authors = [
    {name = "Felix Soubelet", email = "felix.soubelet@github.com"},
]
license = "MIT"
dynamic = ["version"]
requires-python = ">=3.10"

classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Utilities",
]

dependencies = [
    "Pillow >= 10.0",
    "loguru < 1.0",
    "typer >= 0.10",
]

[project.optional-dependencies]
fast = [
    "numba >= 0.60.0",
    "joblib >= 1.4",
]

[project.scripts]
colorbar = "movie_colorbar.__main__:app"

[project.urls]
homepage = "https://github.com/fsoubelet/Movie_Colorbar"
repository = "https://github.com/fsoubelet/Movie_Colorbar"

# ----- Dev Tools Configuration ----- #

[tool.isort]
atomic = true
ensure_newline_before_comments = true
filter_files = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
known_first_party = "movie_colorbar"
line_length = 100
lines_between_types = 1
multi_line_output = 3
use_parentheses = true
profile = "black"

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py310"  # Assume Python 3.10+
